requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL==1.1.0
pandas
requests==2.31.0
pdfkit
numpy
tqdm # Used in get_atlas_data.py
scipy
matplotlib # Used in osm_analysis.py for plotting
sqlalchemy # Used in import_data_db.py for database operations
cryptography

# Auth and security
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode==7.4.2
Flask-Talisman==1.1.0
Flask-Limiter==3.8.0